+++ onsenui/CHANGELOG.md
@@ -24,6 +24,7 @@ dev
  * ons-button, ons-alert-dialog-button, ons-toolbar-button, ons-action-sheet-button: Ensure ons-ripple is imported when button is imported.
  * ons-dialog, ons-alert-dialog, ons-modal, ons-popover, ons-action-sheet: `dialog-cancel` event is renamed to `dialogcancel`.
  * ons-action-sheet: Hide title element when title is null or undefined.
+ * ons-splitter-side: Fix error being thrown when `animation-options` attribute is set.
 
  ### New Features